+++ lib/domain/index.ts
@@ -1,3 +1,4 @@
+export * from "./growth";
 export * from "./measurement";
 export * from "./pet";
 export * from "./units";